firmware_update: Fix str_type encoding size
Cast str_type to uint8_t before inserting into the message buffer
in encode_get_firmware_parameters_resp(). The str_type field is
declared as enum pldm_firmware_update_string_type which is int-sized,
causing pldm_msgbuf_insert's _Generic dispatch to write 4 bytes
instead of the 1 byte expected by the wire format. This corrupts
all subsequent fields in the response, causing decoders to return
PLDM_ERROR_INVALID_DATA.

The component entry encoder already had the correct cast.

+++ src/dsp/firmware_update.c
@@ -1346,12 +1346,14 @@ int encode_get_firmware_parameters_resp(
 	pldm_msgbuf_insert(buf, resp_data->completion_code);
 	pldm_msgbuf_insert(buf, resp_data->capabilities_during_update.value);
 	pldm_msgbuf_insert(buf, resp_data->comp_count);
-	pldm_msgbuf_insert(buf,
-			   resp_data->active_comp_image_set_ver_str.str_type);
+	pldm_msgbuf_insert(
+		buf,
+		(uint8_t)resp_data->active_comp_image_set_ver_str.str_type);
 	pldm_msgbuf_insert(buf,
 			   resp_data->active_comp_image_set_ver_str.str_len);
-	pldm_msgbuf_insert(buf,
-			   resp_data->pending_comp_image_set_ver_str.str_type);
+	pldm_msgbuf_insert(
+		buf,
+		(uint8_t)resp_data->pending_comp_image_set_ver_str.str_type);
 	pldm_msgbuf_insert(buf,
 			   resp_data->pending_comp_image_set_ver_str.str_len);
 	/* String data appended */
